feat: systematize assimilation of cofunction
