+++ src/xfp/functions/__init__.py
@@ -4,6 +4,8 @@ from xfp.functions.xfunc import XFunc
 # fmt: off
 __all__ = [
     "F0", "F1", "F2", "F3", "F4", "F5", "F6", "F7", "F8", "F9", "F10", "F11", "F12", "F13", "F14", "F15", "F16", "F17", "F18", "F19", "F20", "F21", "F22",  # noqa: F405
+    "AF0", "AF1", "AF2", "AF3", "AF4", "AF5", "AF6", "AF7", "AF8", "AF9", "AF10", "AF11", "AF12", "AF13", "AF14", "AF15", "AF16", "AF17", "AF18", "AF19", "AF20", "AF21", "AF22",  # noqa: AF405, F405
+    "XF0", "XF1", "XF2", "XF3", "XF4", "XF5", "XF6", "XF7", "XF8", "XF9", "XF10", "XF11", "XF12", "XF13", "XF14", "XF15", "XF16", "XF17", "XF18", "XF19", "XF20", "XF21", "XF22",  # noqa: F405, XF405
     "curry", "curry0", "curry1", "curry2", "curry3", "curry4", "curry5", "curry6", "curry7", "curry8", "curry9", "curry10", "curry11", "curry12", "curry13", "curry14", "curry15", "curry16", "curry17", "curry18", "curry19", "curry20", "curry21", "curry22",  # noqa: F405
     "curry_method", "curry_method0", "curry_method1", "curry_method2", "curry_method3", "curry_method4", "curry_method5", "curry_method6", "curry_method7", "curry_method8", "curry_method9", "curry_method10", "curry_method11", "curry_method12", "curry_method13", "curry_method14", "curry_method15", "curry_method16", "curry_method17", "curry_method18", "curry_method19", "curry_method20", "curry_method21", "curry_method22",  # noqa: F405
     "tupled", "tupled1", "tupled2", "tupled3", "tupled4", "tupled5", "tupled6", "tupled7", "tupled8", "tupled9", "tupled10", "tupled11", "tupled12", "tupled13", "tupled14", "tupled15", "tupled16", "tupled17", "tupled18", "tupled19", "tupled20", "tupled21", "tupled22",  # noqa: F405
