+++ src/xfp/functions/utils.py
@@ -1,4 +1,4 @@
-from typing import Callable
+from typing import Callable, Coroutine
 from inspect import signature
 from functools import partial
 from typing import TypeVar, ParamSpec, cast, Concatenate, Any
@@ -37,6 +37,58 @@ type F19[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, 
 type F20[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, Out] =                 F1[In1, F19[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, Out]]
 type F21[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, **In21, Out] =         F1[In1, F20[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, Out]]
 type F22[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, **In21, **In22, Out] = F1[In1, F21[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22, Out]]
+# fmt: on
+
+# fmt: off
+type AF0[Out] =                                                                                                                                                                         Callable[[], Coroutine[Any, Any, Out]]
+type AF1[**In, Out] =                                                                                                                                                                   Callable[In, Coroutine[Any, Any, Out]]
+type AF2[**In1, **In2, Out] =                                                                                                                                                           F1[In1, AF1[In2, Out]]
+type AF3[**In1, **In2, **In3, Out] =                                                                                                                                                    F1[In1, AF2[In2, In3, Out]]
+type AF4[**In1, **In2, **In3, **In4, Out] =                                                                                                                                             F1[In1, AF3[In2, In3, In4, Out]]
+type AF5[**In1, **In2, **In3, **In4, **In5, Out] =                                                                                                                                      F1[In1, AF4[In2, In3, In4, In5, Out]]
+type AF6[**In1, **In2, **In3, **In4, **In5, **In6, Out] =                                                                                                                               F1[In1, AF5[In2, In3, In4, In5, In6, Out]]
+type AF7[**In1, **In2, **In3, **In4, **In5, **In6, **In7, Out] =                                                                                                                        F1[In1, AF6[In2, In3, In4, In5, In6, In7, Out]]
+type AF8[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, Out] =                                                                                                                 F1[In1, AF7[In2, In3, In4, In5, In6, In7, In8, Out]]
+type AF9[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, Out] =                                                                                                          F1[In1, AF8[In2, In3, In4, In5, In6, In7, In8, In9, Out]]
+type AF10[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, Out] =                                                                                                 F1[In1, AF9[In2, In3, In4, In5, In6, In7, In8, In9, In10, Out]]
+type AF11[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, Out] =                                                                                         F1[In1, AF10[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, Out]]
+type AF12[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, Out] =                                                                                 F1[In1, AF11[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, Out]]
+type AF13[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, Out] =                                                                         F1[In1, AF12[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, Out]]
+type AF14[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, Out] =                                                                 F1[In1, AF13[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, Out]]
+type AF15[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, Out] =                                                         F1[In1, AF14[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, Out]]
+type AF16[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, Out] =                                                 F1[In1, AF15[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, Out]]
+type AF17[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, Out] =                                         F1[In1, AF16[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, Out]]
+type AF18[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, Out] =                                 F1[In1, AF17[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, Out]]
+type AF19[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, Out] =                         F1[In1, AF18[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, Out]]
+type AF20[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, Out] =                 F1[In1, AF19[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, Out]]
+type AF21[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, **In21, Out] =         F1[In1, AF20[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, Out]]
+type AF22[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, **In21, **In22, Out] = F1[In1, AF21[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22, Out]]
+# fmt: on
+
+# fmt: off
+type XF0[Out] =                                                                                                                                                                         AF0[Out] | F0[Out]
+type XF1[**In, Out] =                                                                                                                                                                   AF1[In, Out] | F1[In, Out]
+type XF2[**In1, **In2, Out] =                                                                                                                                                           AF2[In1, In2, Out] | F2[In1, In2, Out]
+type XF3[**In1, **In2, **In3, Out] =                                                                                                                                                    AF3[In1, In2, In3, Out] | F3[In1, In2, In3, Out]
+type XF4[**In1, **In2, **In3, **In4, Out] =                                                                                                                                             AF4[In1, In2, In3, In4, Out] | F4[In1, In2, In3, In4, Out]
+type XF5[**In1, **In2, **In3, **In4, **In5, Out] =                                                                                                                                      AF5[In1, In2, In3, In4, In5, Out] | F5[In1, In2, In3, In4, In5, Out]
+type XF6[**In1, **In2, **In3, **In4, **In5, **In6, Out] =                                                                                                                               AF6[In1, In2, In3, In4, In5, In6, Out] | F6[In1, In2, In3, In4, In5, In6, Out]
+type XF7[**In1, **In2, **In3, **In4, **In5, **In6, **In7, Out] =                                                                                                                        AF7[In1, In2, In3, In4, In5, In6, In7, Out] | F7[In1, In2, In3, In4, In5, In6, In7, Out]
+type XF8[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, Out] =                                                                                                                 AF8[In1, In2, In3, In4, In5, In6, In7, In8, Out] | F8[In1, In2, In3, In4, In5, In6, In7, In8, Out]
+type XF9[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, Out] =                                                                                                          AF9[In1, In2, In3, In4, In5, In6, In7, In8, In9, Out] | F9[In1, In2, In3, In4, In5, In6, In7, In8, In9, Out]
+type XF10[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, Out] =                                                                                                 AF10[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, Out] | F10[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, Out]
+type XF11[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, Out] =                                                                                         AF11[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, Out] | F11[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, Out]
+type XF12[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, Out] =                                                                                 AF12[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, Out] | F12[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, Out]
+type XF13[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, Out] =                                                                         AF13[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, Out] | F13[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, Out]
+type XF14[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, Out] =                                                                 AF14[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, Out] | F14[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, Out]
+type XF15[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, Out] =                                                         AF15[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, Out] | F15[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, Out]
+type XF16[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, Out] =                                                 AF16[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, Out] | F16[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, Out]
+type XF17[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, Out] =                                         AF17[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, Out] | F17[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, Out]
+type XF18[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, Out] =                                 AF18[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, Out] | F18[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, Out]
+type XF19[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, Out] =                         AF19[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, Out] | F19[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, Out]
+type XF20[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, Out] =                 AF20[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, Out] | F20[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, Out]
+type XF21[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, **In21, Out] =         AF21[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, Out] | F21[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, Out]
+type XF22[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, **In21, **In22, Out] = AF22[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22, Out] | F22[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22, Out]
 # fmt: on
